+++ internal/util/cep_validate.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import (
 	"fmt"
